requirements.txt
streamlit
google-generativeai
python-dotenv
streamlit
pathlib
textwrap